Fix some clang-tidy warnings

+++ src/jngl/SoundFile.cpp
@@ -171,7 +171,7 @@ SoundFile::SoundFile(const std::string& filename, std::launch)
 	}
 
 	internal::debug("Decoded {} ({:.2f} MB, {})", filename,
-	                buffer->size() * sizeof(float) / 1024. / 1024.,
+	                static_cast<double>(buffer->size()) * sizeof(float) / 1024. / 1024.,
 #if !defined(__APPLE__) /* FIXME: Remove when AppleClang's libc++ supports this C++20 feature */   \
     && defined(__GNUC__) && __GNUC__ > 13 // Ubuntu 22.04's GCC doesn't fully support C++20
 	                std::chrono::duration_cast<std::chrono::seconds>(length())
